Add remoteDataManager tests and organize files
