+++ LiveNationTest/LiveNationTestTests/SearchRemoteDataManagerTests.swift
@@ -1,0 +1,8 @@
+//
+//  SearchRemoteDataManagerTests.swift
+//  LiveNationTestTests
+//
+//  Created by Omar Gonzalez on 17/03/24.
+//
+
+import Foundation
